requirements.txt
langchain 
langchain-groq 
langchain-community
langchain-huggingface 
chromadb 
rank_bm25 
python-pptx 
python-dotenv 
unstructured 
networkx 
scikit-learn